# Improve performance by using for-of loop instead of iterate functions