--- a/src/markdown/background_image.js
+++ b/src/markdown/background_image.js
@@ -30,22 +30,20 @@ function backgroundImage(md) {
     'marpit_parse_image',
     'marpit_background_image',
     ({ tokens }) => {
-      tokens.forEach(t => {
-        if (t.type !== 'image') return
-
-        if (t.meta.marpitImage.options.includes('bg')) {
+      for (const t of tokens) {
+        if (t.type === 'image' && t.meta.marpitImage.options.includes('bg')) {
           t.meta.marpitImage.background = true
           t.hidden = true
 
-          t.meta.marpitImage.options.forEach(opt => {
+          for (const opt of t.meta.marpitImage.options) {
             if (bgSizeKeywords[opt])
               t.meta.marpitImage.backgroundSize = bgSizeKeywords[opt]
 
             if (opt === 'left' || opt === 'right')
               t.meta.marpitImage.backgroundSplit = opt
-          })
+          }
         }
-      })
+      }
     }
   )
 
@@ -57,7 +55,7 @@ function backgroundImage(md) {
 
       let current = {}
 
-      tokens.forEach(tb => {
+      for (const tb of tokens) {
         if (tb.type === 'marpit_slide_open') current.open = tb
         if (tb.type === 'marpit_inline_svg_content_open')
           current.svgContent = tb
@@ -92,44 +90,44 @@ function backgroundImage(md) {
           current = {}
         }
 
-        if (!current.open || tb.type !== 'inline') return
-
-        tb.children.forEach(t => {
-          if (t.type !== 'image') return
-
-          const {
-            background,
-            backgroundSize,
-            backgroundSplit,
-            filter,
-            height,
-            size,
-            url,
-            width,
-          } = t.meta.marpitImage
-
-          if (background && !url.match(/^\s*$/)) {
-            current.images = [
-              ...(current.images || []),
-              {
+        if (current.open && tb.type === 'inline')
+          for (const t of tb.children) {
+            if (t.type === 'image') {
+              const {
+                background,
+                backgroundSize,
+                backgroundSplit,
                 filter,
                 height,
-                size: (() => {
-                  const s = size || backgroundSize || undefined
-
-                  return !['contain', 'cover'].includes(s) && (width || height)
-                    ? `${width || s || 'auto'} ${height || s || 'auto'}`
-                    : s
-                })(),
+                size,
                 url,
                 width,
-              },
-            ]
-          }
+              } = t.meta.marpitImage
 
-          if (backgroundSplit) current.split = backgroundSplit
-        })
-      })
+              if (background && !url.match(/^\s*$/)) {
+                current.images = [
+                  ...(current.images || []),
+                  {
+                    filter,
+                    height,
+                    size: (() => {
+                      const s = size || backgroundSize || undefined
+
+                      return !['contain', 'cover'].includes(s) &&
+                        (width || height)
+                        ? `${width || s || 'auto'} ${height || s || 'auto'}`
+                        : s
+                    })(),
+                    url,
+                    width,
+                  },
+                ]
+              }
+
+              if (backgroundSplit) current.split = backgroundSplit
+            }
+          }
+      }
     }
   )
 
@@ -139,6 +137,7 @@ function backgroundImage(md) {
     state => {
       let current
 
+      // TODO: Use for-of loop isntead of reduce
       state.tokens = state.tokens.reduce((ret, t) => {
         let tokens = [t]