Use `for-of` loop instead of iterate functions like `Array.forEach()` or `Array.reduce()`. It receives code readability and better performance.

## Benchmark

We have measured the time of 1000 conversions from [example of uncover theme](https://gist.github.com/yhatt/a352e6087c2d5185113b9143f330fb68). Averages of 20 trials are below:

||Node v8|Node v6|
|:---:|:---:|:---:|
|**Iterate functions**|2162.35ms|2787.40ms|
|**for-of**|1227.90ms|1954.70ms|

```javascript
const request = require('request')
const { Marpit } = require('./lib/')

const average = arr => arr.reduce((p, c) => p + c, 0) / arr.length
const marpit = new Marpit()

request.get(
  'https://gist.githubusercontent.com/yhatt/a352e6087c2d5185113b9143f330fb68/raw/ca3ca25b3a022ccaae3836c0464b02693230dcef/_uncover.md',
  (_, __, body) => {
    const tries = []

    for (let t = 0; t < 20; t += 1) {
      const start = new Date().getTime()

      for (let i = 0; i < 1000; i += 1) marpit.render(body)
      const end = new Date().getTime() - start

      tries.push(end)
      console.log(`Try ${t + 1}: ${end}ms`)
    }

    console.log(`Average: ${average(tries)}ms`)
  }
)
```