@yhatt 2018-10-05
UPDATE: Add 5 times idling in benchmark code. We have re-measured and it seems to be about 40-70% faster.